+++ src/Item/index.js
@@ -1,1 +1,3 @@
 export * from './CreateItem';
+export * from './DuplicateItem';
+export * from './EditItem';