+++ src/index.js
@@ -32,10 +32,11 @@ import {
   ViewRequestsRoute,
   ImportRoute,
   HoldingsMarcRoute,
+  EditItemRoute,
+  DuplicateItemRoute,
 } from './routes';
 import Settings from './settings';
 import { DataProvider, HoldingsProvider } from './providers';
-import { commands } from './constants';
 
 const InventoryRouting = (props) => {
   const [isShortcutsModalOpen, setIsShortcutsModalOpen] = useState(false);
@@ -98,6 +99,10 @@ const InventoryRouting = (props) => {
                 component={CreateHoldingRoute}
               />
               <Route
+                path={`${path}/edit/:id/:holdingId/:itemId`}
+                component={EditItemRoute}
+              />
+              <Route
                 path={`${path}/create/:id/:holdingId/item`}
                 component={CreateItemRoute}
               />
@@ -108,6 +113,10 @@ const InventoryRouting = (props) => {
               <Route
                 path={`${path}/view/:id/:holdingsrecordid/:itemid`}
                 component={ItemRoute}
+              />
+              <Route
+                path={`${path}/copy/:id/:holdingsrecordid/:itemid`}
+                component={DuplicateItemRoute}
               />
               <Route
                 path={`${path}/quick-marc`}
@@ -146,7 +155,7 @@ const InventoryRouting = (props) => {
         </CommandList>
         {isShortcutsModalOpen && (
           <KeyboardShortcutsModal
-            allCommands={[...defaultKeyboardShortcuts, ...commands]}
+            allCommands={defaultKeyboardShortcuts}
             onClose={toggleModal}
           />
         )}