+++ src/views/ItemView.js
@@ -1,7 +1,6 @@
 import {
   get,
   cloneDeep,
-  omit,
   map,
   isEmpty,
   values,
@@ -10,7 +9,6 @@ import { parameterize } from 'inflected';
 
 import React, { createRef } from 'react';
 import PropTypes from 'prop-types';
-import queryString from 'query-string';
 import { Link } from 'react-router-dom';
 import { FormattedMessage } from 'react-intl';
 
@@ -27,7 +25,6 @@ import {
   AccordionStatus,
   ExpandAllButton,
   KeyValue,
-  Layer,
   MultiColumnList,
   Button,
   Icon,
@@ -70,7 +67,6 @@ import {
   checkIfArrayIsEmpty,
   handleKeyCommand,
 } from '../utils';
-import ItemForm from '../edit/items/ItemForm';
 import withLocation from '../withLocation';
 import {
   itemStatusesMap,
@@ -105,7 +101,13 @@ class ItemView extends React.Component {
 
   onClickEditItem = e => {
     if (e) e.preventDefault();
-    this.props.updateLocation({ layer: 'editItem' });
+
+    const { id, holdingsrecordid, itemid } = this.props.match.params;
+
+    this.props.history.push({
+      pathname: `/inventory/edit/${id}/${holdingsrecordid}/${itemid}`,
+      search: location.search,
+    });
   };
 
   onClickCloseEditItem = e => {
@@ -145,17 +147,13 @@ class ItemView extends React.Component {
     this.props.mutator.items.DELETE(item);
   };
 
-  onCopy(item) {
-    this.setState((state) => {
-      const newState = cloneDeep(state);
+  onCopy() {
+    const { itemid, id, holdingsrecordid } = this.props.match.params;
 
-      newState.copiedItem = omit(item, ['id', 'hrid', 'barcode']);
-      newState.copiedItem.status = { name: 'Available' };
-
-      return newState;
+    this.props.history.push({
+      pathname: `/inventory/copy/${id}/${holdingsrecordid}/${itemid}`,
+      search: location.search,
     });
-
-    this.props.updateLocation({ layer: 'copyItem' });
   }
 
   markItemAsMissing = () => {
@@ -394,7 +392,6 @@ class ItemView extends React.Component {
 
   render() {
     const {
-      location,
       resources: {
         items,
         holdingsRecords,
@@ -406,7 +403,6 @@ class ItemView extends React.Component {
         tagSettings,
       },
       referenceTables,
-      okapi,
       goTo,
       stripes,
     } = this.props;
@@ -417,7 +413,6 @@ class ItemView extends React.Component {
       itemMissingModal,
       itemWithdrawnModal,
       selectedItemStatus,
-      copiedItem,
       confirmDeleteItemModal,
     } = this.state;
 
@@ -443,7 +438,6 @@ class ItemView extends React.Component {
     const tagsEnabled = !tagSettings?.records?.length || tagSettings?.records?.[0]?.value === 'true';
 
     const requestCount = requests.other?.totalRecords ?? 0;
-    const query = location.search ? queryString.parse(location.search) : {};
 
     const requestsUrl = `/requests?filters=${requestStatusFiltersString}&query=${item.id}&sort=Request Date`;
 
@@ -717,6 +711,12 @@ class ItemView extends React.Component {
       {
         name: 'search',
         handler: handleKeyCommand(() => goTo('/inventory')),
+      },
+      {
+        name: 'duplicateRecord',
+        handler: handleKeyCommand(() => {
+          if (stripes.hasPerm('ui-inventory.item.create')) this.onCopy();
+        }),
       },
     ];
 
@@ -1435,39 +1435,6 @@ class ItemView extends React.Component {
                   </AccordionSet>
                 </AccordionStatus>
               </Pane>
-              <Layer
-                isOpen={query.layer ? query.layer === 'editItem' : false}
-                contentLabel={intl.formatMessage({ id: 'ui-inventory.editItemDialog' })}
-              >
-                <ItemForm
-                  form={`itemform_${item.id}`}
-                  onSubmit={record => this.saveItem(record)}
-                  initialValues={item}
-                  onCancel={this.onClickCloseEditItem}
-                  okapi={okapi}
-                  instance={instance}
-                  holdingsRecord={holdingsRecord}
-                  referenceTables={referenceTables}
-                  stripes={this.props.stripes}
-                />
-              </Layer>
-              <Layer
-                isOpen={query.layer === 'copyItem'}
-                contentLabel={intl.formatMessage({ id: 'ui-inventory.copyItemDialog' })}
-              >
-                <ItemForm
-                  form={`itemform_${holdingsRecord.id}`}
-                  onSubmit={(record) => { this.copyItem(record); }}
-                  initialValues={copiedItem}
-                  onCancel={this.onClickCloseEditItem}
-                  okapi={okapi}
-                  instance={instance}
-                  copy
-                  holdingsRecord={holdingsRecord}
-                  referenceTables={referenceTables}
-                  stripes={this.props.stripes}
-                />
-              </Layer>
             </Paneset>
           </HasCommand>
         )}
@@ -1539,6 +1506,8 @@ ItemView.propTypes = {
   onCloseViewItem: PropTypes.func.isRequired,
   updateLocation: PropTypes.func.isRequired,
   goTo: PropTypes.func.isRequired,
+  match: PropTypes.object.isRequired,
+  history: PropTypes.object.isRequired,
 };
 
 export default withLocation(ItemView);